# openapi.yaml
openapi: 3.1.0
info:
  title: Hermes Agents WebApp API
  version: 1.0.0
  description: File-system based Hermes agent management API
servers:
  - url: http://localhost:18470
    description: Local development
  - url: https://hermes-agents.mini.tumf.dev
    description: Production (mini)
paths:
  /api/agents:
    get:
      summary: List all agents
      operationId: listAgents
      tags: [Agents]
      responses:
        '200':
          description: Agent list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Agent'
    post:
      summary: Create a new agent
      operationId: createAgent
      tags: [Agents]
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                templates:
                  type: object
                  properties:
                    memoryMd:
                      type: string
                    userMd:
                      type: string
                    soulMd:
                      type: string
                    configYaml:
                      type: string
                meta:
                  $ref: '#/components/schemas/AgentMeta'
      responses:
        '201':
          description: Created agent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Agent'
    delete:
      summary: Delete an agent
      operationId: deleteAgent
      tags: [Agents]
      parameters:
        - name: id
          in: query
          required: true
          schema:
            type: string
        - name: purge
          in: query
          schema:
            type: string
            enum: ['true']
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
...
  /api/files:
    get:
      summary: Get agent file content
      operationId: getFile
      tags: [Files]
      parameters:
        - name: agent
          in: query
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
            enum: [MEMORY.md, USER.md, SOUL.md, config.yaml]
      responses:
        '200':
          description: File content
          content:
            application/json:
              schema:
                type: object
                properties:
                  content:
                    type: string
    put:
      summary: Update agent file content
      operationId: putFile
      tags: [Files]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [agent, path, content]
              properties:
                agent:
                  type: string
                path:
                  type: string
                  enum: [MEMORY.md, USER.md, SOUL.md, config.yaml]
                content:
                  type: string
      responses:
        '200':
          description: Saved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
        '422':
          description: Invalid YAML (config.yaml only)
...
  /api/templates:
    get:
      summary: List templates or get file content
      operationId: getTemplates
      tags: [Templates]
      parameters:
        - name: name
          in: query
          schema:
            type: string
        - name: file
          in: query
          schema:
            type: string
            enum: [MEMORY.md, USER.md, SOUL.md, config.yaml]
      responses:
        '200':
          description: Template list or file content
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items:
                      $ref: '#/components/schemas/Template'
                  - type: object
                    properties:
                      name:
                        type: string
                      file:
                        type: string
                      content:
                        type: string
    post:
      summary: Create a template file
      operationId: createTemplate
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateFile'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateFile'
        '409':
          description: Already exists
    put:
      summary: Update a template file
      operationId: updateTemplate
      tags: [Templates]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TemplateFile'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TemplateFile'
        '404':
          description: Not found
    delete:
      summary: Delete a template or template file
      operationId: deleteTemplate
      tags: [Templates]
      parameters:
        - name: name
          in: query
          required: true
          schema:
            type: string
        - name: file
          in: query
          schema:
            type: string
            enum: [MEMORY.md, USER.md, SOUL.md, config.yaml]
      responses:
        '200':
          description: Deleted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OkResponse'
components:
  schemas:
    OkResponse:
      type: object
      properties:
        ok:
          type: boolean
          const: true
    Agent:
      type: object
      properties:
        agentId:
          type: string
        home:
          type: string
        label:
          type: string
        enabled:
          type: boolean
        createdAt:
          type: string
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
    AgentMeta:
      type: object
      properties:
        name:
          type: string
        description:
          type: string
        tags:
          type: array
          items:
            type: string
...
    Template:
      type: object
      properties:
        name:
          type: string
        files:
          type: array
          items:
            type: string
    TemplateFile:
      type: object
      required: [name, file, content]
      properties:
        name:
          type: string
        file:
          type: string
          enum: [MEMORY.md, USER.md, SOUL.md, config.yaml]
        content:
          type: string